refactor: rename util name
