+++ app/components/Pages/CreateAccount/UserPrevetForm/index.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faChevronRight } from '@fortawesome/free-solid-svg-icons';
-import { validateUserEmail } from 'utils/validate-form-inputs';
+import { validateUserEmail } from 'utils/validateFormInputs';
 import { FormTitle, InputAndButtonWrapper } from '../CreateAccount.styles';
 import FloatingLabelInput from '../FloatingLabelInput';
 import Button from 'components/Button';
